Renaming packages to make tests more obvious
